requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
pandas==2.1.4
openai==1.3.7
google-cloud-texttospeech==2.16.3
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.25.2